Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -3,12 +3,13 @@ package com.educandoweb.course.services;
 import java.util.List;
 import java.util.Optional;
 
-import org.hibernate.ResourceClosedException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exceptions.DatabaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,8 +31,20 @@ public class UserService {
 		return repository.save(obj);
 	}
 	
-	public void delete(Long id) {
-		repository.deleteById(id);
+	public void delete(Long id){
+	    try {
+	        if(!repository.existsById(id)) {
+	        	throw new ResourceNotFoundException(id);
+	        }
+	        
+	        repository.deleteById(id);
+	    }
+	    catch (ResourceNotFoundException e){
+	        throw new ResourceNotFoundException(id);
+	    }
+	    catch(DataIntegrityViolationException e) {
+	    	throw new DatabaseException(e.getMessage());
+	    }
 	}
 	
 	public User update(Long id, User obj) {
